% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxspell-package.R
\docType{package}
\name{taxspell-package}
\alias{taxspell-package}
\alias{taxspell}
\title{taxspell}
\description{
Spell Check Taxonomic Names
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
