% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{ts_cache}
\alias{ts_cache}
\title{Caching}
\description{
Manage cached \code{taxspell} files with \pkg{hoardr}
}
\details{
The dafault cache directory is
\code{paste0(rappdirs::user_cache_dir(), "/R/taxspell")}, but you can set
your own path using \code{cache_path_set()}

\code{cache_delete} only accepts 1 file name, while
\code{cache_delete_all} doesn't accept any names, but deletes all files.
For deleting many specific files, use \code{cache_delete} in a \code{\link[=lapply]{lapply()}}
type call
}
\section{Useful user functions}{

\itemize{
\item \code{ts_cache$cache_path_get()} get cache path
\item \code{ts_cache$cache_path_set()} set cache path. You can set the
entire path directly via the \code{full_path} arg like
\code{ts_cache$cache_path_set(full_path = "your/path")}
\item \code{ts_cache$list()} returns a character vector of full
path file names
\item \code{ts_cache$files()} returns file objects with metadata
\item \code{ts_cache$details()} returns files with details
\item \code{ts_cache$delete()} delete specific files
\item \code{ts_cache$delete_all()} delete all files, returns nothing
}
}

\examples{
\dontrun{
ts_cache

# list files in cache
ts_cache$list()

# delete certain database files
# ts_cache$delete("file path")
# ts_cache$list()

# delete all files in cache
# ts_cache$delete_all()
# ts_cache$list()
}
}
