% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_check.R
\name{ts_check}
\alias{ts_check}
\title{Check taxonomic spelling}
\usage{
ts_check(x, dict = ts_dict("plants_tpl"))
}
\arguments{
\item{x}{(character) vector of taxonomic names. required}

\item{dict}{(hunspell_dictionary) a hunspell dictionary, see \code{\link[=ts_dict]{ts_dict()}}}
}
\value{
tibble/data.frame of the form:
\itemize{
\item supplied (character): the supplied name in the \code{x} parameter
\item match (logical): whether an exact match was found
\item matches (character): the matched named, if an exact match was found
\item suggestions (list): if no match found, a vector of suggestions
}
}
\description{
Check taxonomic spelling
}
\details{
uninomials only for now
}
\examples{
\dontrun{
x <- c("Helianthus", "Helianthos", "Helionthus",
"Helianthuss", "helianthus", "Hellianthos")
(z <- ts_check(x))
z$suggestions
}
}
