% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dict.R
\name{ts_dict}
\alias{ts_dict}
\title{Check taxonomic spelling}
\usage{
ts_dict(x = "tpl")
}
\arguments{
\item{x}{(character) dictionary to use. options:
"tpl" (The Plant List), "taxamatch_ref" (TAXAMATCH
reference dictionary). default: "tpl".}
}
\value{
path to location of the dictionary files
}
\description{
Check taxonomic spelling
}
\examples{
\dontrun{
ts_dict()
}
}
