% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_fetch_dictionary.R
\name{ts_fetch_dictionary}
\alias{ts_fetch_dictionary}
\title{download a dictionary}
\usage{
ts_fetch_dictionary(x, ...)
}
\arguments{
\item{x}{(character) name of a dictionary to download. options:
"tpl" (The Plant List), "taxamatch_ref" (TAXAMATCH reference dictionary)}

\item{...}{arguments passed down to \code{\link[=download.file]{download.file()}}}
}
\value{
nothing, downloads dictionary to the cache path, see \link{ts_cache}
}
\description{
download a dictionary
}
\examples{
# ts_fetch_dictionary("tpl")
}
