% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ts_make_dictionary}
\alias{ts_make_dictionary}
\title{make a dictionary}
\usage{
ts_make_dictionary(x, name = NULL, path = NULL)
}
\arguments{
\item{x}{(character) vector of taxonomic names. required. if names
are given with spaces (e.g., a species name like \emph{Poa annua}) we'll
split those up, with each written to a new line}

\item{name}{(character) name to use for the dictionary files}

\item{path}{(character) path in which to put dictionary files. required}
}
\description{
make a dictionary
}
\details{
if \code{name} is \code{NULL}, the name of the dictionary files is taken
from the file path
}
\examples{
# ts_make_dictionary("x", name = "foo")
}
\keyword{internal}
