% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dict.R
\name{ts_dict}
\alias{ts_dict}
\title{Check taxonomic spelling}
\usage{
ts_dict(x = "plants_tpl")
}
\arguments{
\item{x}{(character) dictionary to use. default: "plants_tpl".
other options will be added, including custom ones}
}
\value{
path to location of the dictionary files
}
\description{
Check taxonomic spelling
}
\examples{
ts_dict()
}
